Panaversity Class Learning/intro_tool_calling/pyproject.toml
[project]
name = "intro-tool-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.39.1",
    "openai-agents>=0.3.3",
]
